api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
psycopg[binary]==3.2.3
python-jose[cryptography]==3.3.0
# bcrypt>=4 removed __about__.__version__, which some stacks still expect.
bcrypt==3.2.2
email-validator==2.2.0
httpx==0.27.2
river==0.21.2
# Drivers para ingestión desde bases de datos externas
PyMySQL==1.1.1
pyodbc==5.2.0
